requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
gspread>=5.0.0
google-auth>=2.3.0
apscheduler>=3.9.0
flask>=2.0.0